@@ -8,11 +8,15 @@ namespace E_Commerce.CatalogService.Application.Abstractions.Services
     public interface ICatalogItemsService
     {
         Task CreateProduct(CreateCatalogItemDto createCatalogItemDto, CreateCatalogItemFeaturesDto createCatalogItemFeaturesDto);
-        Task<CatalogItem> GetByIdWithImage(int CatalogItemId);
+        Task<CatalogItem> GetByIdWithImage(uint CatalogItemId);
         Task UpdateProduct(UpdateCatalogItemDto updateCatalogItemDto);
 
-        Task<GetCatalogItemDto> GetCatalogItem(int CatalogItemId);
+        Task<GetCatalogItemDto> GetCatalogItem(uint CatalogItemId);
 
         Task<CatalogItemListModel> GetListCatalogItem(PageRequest pageRequest);
+
+        Task<CatalogItemListModel?> GetListCatalogItemByType(PageRequest pageRequest, uint catalogTypeId);
+
+        Task<CatalogItemListModel?> GetListCatalogItemByBrand(PageRequest pageRequest, uint brandId);
     }
 }
