@@ -1,4 +1,7 @@
-﻿using MediatR;
+﻿using E_Commerce.CatalogService.Application.Features.CatalogBrands.Rules;
+using E_Commerce.CatalogService.Application.Features.CatalogCategory.Rule;
+using E_Commerce.CatalogService.Application.Features.CatalogItems.Rules;
+using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 
@@ -10,6 +13,9 @@ namespace E_Commerce.CatalogService.Application.Extensions
         {
             services.AddMediatR(Assembly.GetExecutingAssembly());
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
+            services.AddScoped<CatalogItemBusinessRule>();
+            services.AddScoped<CatalogTypeBusinessRules>();
+            services.AddScoped<CatalogBrandBusinessRules>();
         }
     }
 }
