@@ -2,7 +2,7 @@
 {
     public class GetListCatalogItemDto
     {
-        public int Id { get; set; }
+        public uint Id { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
         public string CatalogBrandName { get; set; }
