@@ -2,7 +2,7 @@
 {
     public class UpdateCatalogItemDto
     {
-        public int Id { get; set; }
+        public uint Id { get; set; }
         public CreateCatalogItemDto createCatalogItemDto { get; set; }
         public CreateCatalogItemFeaturesDto createCatalogItemFeaturesDto { get; set; }
     }
