@@ -1,6 +1,6 @@
 ﻿namespace E_Commerce.CatalogService.Domain.Entity.Base
 {
-    public class BaseEntity<T> : IEntity
+    public abstract class BaseEntity<T> : IEntity
     {
         public T Id { get; set; }
         public DateTime CreatedDate { get; set; }
