@@ -2,7 +2,7 @@
 
 namespace E_Commerce.CatalogService.Domain.Entity
 {
-    public class CatalogBrand : BaseEntity<ushort>
+    public class CatalogBrand : BaseEntity<uint>
     {
         public string Brand { get; set; }
         public ICollection<CatalogItem> CatalogItems { get; set; }
