@@ -1,6 +1,5 @@
 ﻿using E_Commerce.CatalogService.Domain.Entity;
 using E_Commerce.CatalogService.Domain.enums;
-using E_Commerce.CatalogService.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -10,7 +9,7 @@ namespace E_Commerce.CatalogService.Persistence.EntityConfiguration
     {
         public void Configure(EntityTypeBuilder<CatalogItemFeature> builder)
         {
-            builder.ToTable("Catalog_Item_Feature", CatalogContext.DEFAULT_SCHEMA);
+            builder.ToTable("Catalog_Item_Feature");
             builder.HasKey(x => x.Id);
             builder.Property(x => x.CreatedDate).IsRequired();
             builder.Property(x => x.UpdatedDate).IsRequired(false);
