@@ -1,5 +1,4 @@
 ﻿using E_Commerce.CatalogService.Domain.Entity;
-using E_Commerce.CatalogService.Persistence.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -9,7 +8,7 @@ namespace E_Commerce.CatalogService.Persistence.EntityConfiguration
     {
         public void Configure(EntityTypeBuilder<CatalogItemImage> builder)
         {
-            builder.ToTable("Catalog_Item_Image", CatalogContext.DEFAULT_SCHEMA);
+            builder.ToTable("Catalog_Item_Image");
             builder.HasKey(x => x.Id);
             builder.Property(x => x.CreatedDate).IsRequired();
             builder.Property(x => x.UpdatedDate).IsRequired(false);
