@@ -8,7 +8,7 @@ namespace E_Commerce.OrderService.Application.Extensions
     {
         public static void AddApplicationServiceRegistration(this IServiceCollection services)
         {
-            services.AddMediatR(Assembly.GetExecutingAssembly());
+            services.AddMediatR(cfg => cfg.AsTransient(), Assembly.GetExecutingAssembly());
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
         }
     }
