@@ -5,9 +5,9 @@ using E_Commerce.OrderService.Persistence.Contexts;
 
 namespace E_Commerce.OrderService.Persistence.Concrete.Repository
 {
-    public class BuyerRepository : GenericRepoitory<Buyer, OrderContext>, IBuyerRepository
+    public class BuyerRepository : GenericRepoitory<Buyer>, IBuyerRepository
     {
-        public BuyerRepository(OrderContext context) : base(context)
+        public BuyerRepository(OrderDbContext context) : base(context)
         {
         }
     }
