@@ -5,17 +5,17 @@ using E_Commerce.OrderService.Persistence.Contexts;
 
 namespace E_Commerce.OrderService.Persistence.Concrete.Repository
 {
-    public class OrderRepository : GenericRepoitory<Order, OrderContext>, IOrderRepository
+    public class OrderRepository : GenericRepoitory<Order>, IOrderRepository
     {
-        private readonly OrderContext _orderContext;
-        public OrderRepository(OrderContext context) : base(context)
+        private readonly OrderDbContext _orderContext;
+        public OrderRepository(OrderDbContext context) : base(context)
         {
             _orderContext = context;
         }
 
         public async Task<Order?> GetById(Guid id)
         {
-            var entity = await base.GetAsync(x => x.Id == id);
+            var entity = await GetAsync(x => x.Id == id);
             return entity ??= _orderContext.Orders.Local.FirstOrDefault(i => i.Id == id);
 
         }
