@@ -1,4 +1,5 @@
 ﻿using E_Commerce.OrderService.Application.Abstractions.Repostories;
+using E_Commerce.OrderService.Application.Abstractions.Services;
 using E_Commerce.OrderService.Persistence.Concrete.Repository;
 using E_Commerce.OrderService.Persistence.Contexts;
 using Microsoft.EntityFrameworkCore;
@@ -11,9 +12,15 @@ namespace E_Commerce.OrderService.Persistence.Extensions
     {
         public static void AddPersistenceServiceRegistration(this IServiceCollection services, IConfiguration configuration)
         {
-            services.AddDbContext<OrderContext>(opt => opt.UseSqlServer(configuration.GetConnectionString("MSS")));
+            services.AddDbContext<OrderDbContext>(opt =>
+            {
+                opt.UseSqlServer(configuration.GetConnectionString("MSS"));
+                opt.EnableSensitiveDataLogging();
+            });
+
             services.AddScoped<IBuyerRepository, BuyerRepository>();
             services.AddScoped<IOrderRepository, OrderRepository>();
+            services.AddScoped<IOrderService, Concrete.Services.OrderService>();
         }
     }
 }
