@@ -1,33 +1,38 @@
-using E_Commerce.EventBus.Base.EventBus.Base;
-using E_Commerce.EventBus.Factory;
+using E_Commerce.OrderService.Api.Extensions;
 using E_Commerce.OrderService.Application.Extensions;
 using E_Commerce.OrderService.Persistence.Extensions;
+using Microsoft.OpenApi.Models;
 
 var builder = WebApplication.CreateBuilder(args);
 
+builder.Host.UseDefaultServiceProvider((context, options) =>
+{
+    options.ValidateOnBuild = false;
+    options.ValidateScopes = false;
+});
 // Add services to the container.
+
 
 builder.Services.AddControllers();
 builder.Services.AddApplicationServiceRegistration();
 builder.Services.AddPersistenceServiceRegistration(builder.Configuration);
+builder.Services.AddAuthServices(builder.Configuration);
+builder.Services.ConfigureEventHandlers();
+builder.Services.AddLogging(cfg => cfg.AddConsole());
+builder.Services.ConfigureConsul(builder.Configuration);
+builder.Services.AddSingleton(sp => sp.EventBusRegister());
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSingleton(sp =>
+
+
+builder.Services.AddSwaggerGen(c =>
 {
-    EventBusConfig eventBusConfig = new()
-    {
-        ConnectionRetryCount = 5,
-        EventNameSuffix = "IntegrationEvent",
-        SubscriberClientAppName = "NotificationService",
-        EventBusType = EventBusType.RabbitMQ
-    };
-    return EventBusFactory.Create(eventBusConfig, sp);
+    c.SwaggerDoc("v1", new OpenApiInfo { Title = "OrderService.Api", Version = "v1" });
 });
 
-
-builder.Services.AddSwaggerGen();
-
 var app = builder.Build();
+
+
 
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
@@ -37,9 +42,12 @@ if (app.Environment.IsDevelopment())
 }
 
 app.UseHttpsRedirection();
-
+app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapControllers();
 
-app.Run();
+app.EventConfig();
+app.Start();
+app.RegisterWithConsul(app.Lifetime);
+app.WaitForShutdown();
